tech(TTB-060): Moved game related to components to new directory
