+++ client/src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import Game from './game/Game';
-import GameRobot from './game/GameRobot';
+import Game from './components/Game/Game';
+import GameRobot from './components/Game/GameRobot';
 import Home from './components/Home/Home';
 import Options from './components/Options/Options';
 import Header from './components/Header/Header';
